requirements.txt
# File monitoring
pathtools==0.1.2
watchdog==0.10.2

# GPS
pyserial==3.4
pynmea2==1.15.0

# Thermal image analysis
flirimageextractor==1.4.0
loguru==0.4.1
numpy
Pillow
matplotlib

# Messaging
validate_email
py3DNS
wrapt-timeout-decorator==1.3.1
